21/05/28_endText_Redner
게임 종료시 해당 텍스트 출력되도록 설정

+++ 2Dgame_MetalSlugAttack/2Dgame_MetalSlugAttack/UiManager.h
@@ -17,8 +17,8 @@ typedef struct FrameUi
 	int selectNum;
 	int unitCoast;
 	int calcUnitCoast[3];
-	int currCoolTime;
-	int purchaseCoolTime;
+	float currCoolTime;
+	float purchaseCoolTime;
 	float checkCoolTime;
 };
 typedef struct enemyAp
